+++ skeleton/application/config/hooks.php
@@ -8,7 +8,7 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 | This file lets you define "hooks" to extend CI without hacking the core
 | files.  Please see the user guide for info:
 |
-|	https://codeigniter.com/user_guide/general/hooks.html
+|	https://codeigniter.com/userguide3/general/hooks.html
 |
 */
 use \X\Annotation\AnnotationReader;
@@ -18,16 +18,16 @@ $hook['post_controller_constructor'] = function() {
   if (is_cli())
     return;
   $CI =& get_instance();
-  $accessibility = AnnotationReader::getAccessibility($CI->router->class, $CI->router->method);
+  $meta = AnnotationReader::getAccessibility($CI->router->class, $CI->router->method);
   $isLogin = !empty($_SESSION[SESSION_NAME]);
   $currentPath = lcfirst($CI->router->directory ?? '') . lcfirst($CI->router->class) . '/' . $CI->router->method;
   $defaultPath = '/users/index';
-  $allowRoles = !empty($accessibility->allow_role) ? array_map('trim', explode(',', $accessibility->allow_role)) : null;
-  if (!$accessibility->allow_http)
+  $allowRoles = !empty($meta->allow_role) ? array_map('trim', explode(',', $meta->allow_role)) : null;
+  if (!$meta->allow_http)
     throw new \RuntimeException('HTTP access is not allowed');
-  else if ($isLogin && !$accessibility->allow_login)
+  else if ($isLogin && !$meta->allow_login)
     redirect($defaultPath);
-  else if (!$isLogin && !$accessibility->allow_logoff)
+  else if (!$isLogin && !$meta->allow_logoff)
     redirect('/users/login');
   else if ($isLogin && !empty($allowRoles)) {
     $role = $_SESSION[SESSION_NAME]['role'] ?? '';